minor bugfix to avoid negative time stamps and ASR synchronization problems
git-svn-id: https://semaine.opendfki.de/repos/trunk@256 2252c265-7fd8-49a5-b10e-6a9b7765d42e

+++ c++/src/tum/asr/src/ASR.h
@@ -84,7 +84,7 @@ private:
 	int speakingIndex,seqIdx;
 	int nFeaturesSelected;
 	int *featureIndex;
-	long curTime;
+	long long curTime;
 	int wasSpeaking;
 
 	float prevSpeakingIndex;
